Version 0.9
Completed everything except for CSS styling

- Converted to FFMPEG usage from webp.wasm
- Added framerate and size radios
- Added custom cut-in import
- Removed in-game functionality, completely web-based now
- Added template available for download

+++ Assets/Scripts/WebCommunication/JsonCompiler.cs
@@ -17,13 +17,13 @@ public class JsonCompiler : MonoBehaviour
 
     public class RenderInformationWrapper
     {
-        public int fps;
-        public int totalFrames;
+        public float fps;
+        public float totalFrames;
         public int width;
         public int height;
         public string webpName;
     }
-    public static string CompileWebpInformation(int fps, int frames, int width, int height, string webpName)
+    public static string CompileWebpInformation(float fps, float frames, int width, int height, string webpName)
     {
         RenderInformationWrapper wrapper = new()
         {
